Bookmarks
Menu de favoritos

Mozilla Firefox

Get Help
Customize Firefox
Get Involved
About Us
Pedir ajuda
Personalizar Firefox
Envolva-se
Quem somos
Ajuda e tutoriais

Mozilla Firefox

Ajuda e tutoriais
Personalizar Firefox
Envolva-se
Quem somos
Personalizar Firefox

Ubuntu and Free Software links

Ubuntu
Ubuntu Wiki (community-edited website)
Make a Support Request to the Ubuntu Community
Debian (Ubuntu is based on Debian)

Help and Tutorials
Customize Firefox
Get Involved
About Us
Ubuntu
Ubuntu Wiki (community-edited website)
Make a Support Request to the Ubuntu Community
Debian (Ubuntu is based on Debian)
Apartamento aconchegante no coração de SP - Apartamentos para Alugar em Paraíso
Grupo ofertas de notebook no WhatsApp - Descontos
Corpo Docente - Uningá
Moodle Educ: Criar uma Conta de Usuário Administrador no Moodle
(7) Tutorial R | Gráficos no ggplot2 sem usar código - YouTube
Etiquetas recentes

Importado do Internet Explorer

Links for Brasil

Microsoft Brasil
Microsoft Brasil
MSN Brasil
Windows Brasil

MSN

Sites da Microsoft na Web

Marketplace
Microsoft Brasil
Ofertas da Microsoft
Marketplace
Site do IE na Microsoft.com
Ofertas da Microsoft

Windows Live

Windows Live Mail

MSN

MSN Automóvel
MSN Entretenimento
MSN Esportes
MSN Notícias
MSN Seu Dinheiro
MSN

MSN

MSN Automóvel
MSN Entretenimento
MSN Esportes
MSN Notícias
MSN Seu Dinheiro
MSN

Windows Live

Obtenha o Windows Live
Windows Live Gallery
Windows Live Mail
Windows Live Spaces
Obtenha o Windows Live
Windows Live Mail
Windows Live Spaces

Sites da Microsoft na Web

Marketplace
Ofertas da Microsoft
Site do IE na Microsoft.com
Microsoft Brasil
Microsoft Brasil

Links for Brasil

MSN Brasil
Microsoft Brasil
MSN Brasil

Sites da Microsoft na Web

Links for Brasil

Guia de operações básicas do ScanSnap iX500 (Windows)

Importado do Internet Explorer

Windows Live

Windows Live Gallery

Importado do Internet Explorer

Caixa de entrada - ecacarva@uem.br - E-mail de Universidade Estadual de Maringá
Internet:::B.aNk.ing CA.IXA
I_nt ernet Ban-King....CAIXA
Pergunta · Metabase
Suicide Rates by State
(82) Gráficos de barras para dados categóricos no R - YouTube
Rascunhos (190) - ecacarva@gmail.com - Gmail
(2) WhatsApp
Meu Drive - Google Drive
Qlik Sense Hub
Caixa de entrada - aleksmontanha@gmail.com - Gmail
Qlik Sense Hub
tinyurl.com/LINKSMARTPHONES
Qlik Sense Hub

Barra de favoritos

GPT

AssGovBr
CAJU
eP
Hub
QMC
QV-New
EC
PIÁ
RU
TDE

Férias

Casa Frente Mar 4 Quartos 3wc 12 Pessoas - Casas para Alugar em Tabuleiro, Barra Velha, Santa Catarina, Brasil - Airbnb
Casa ampla a 30m do mar - Casas de férias para Alugar em Itapoá, Santa Catarina, Brasil - Airbnb

STR

Disney+
Star+
Mercado Livre

DES

NPD

Redmine
NPD
GitLab - Groups
GitLab Dashboard Projects
Dicionario
Classificação de Dados da LGPD - Planilhas Google
Repositorios-codigo-fonte-NPD - Documentos Google
Analistas e Sistemas.xlsx - Planilhas Google
Sistema de Publicação de Legislação
Apostila
Creative Tim
NPD-Reestruturação: Lucidchart
prisma | DB Designer
Bizagi Accounts
Desenvolvimento - Documentos Google - Alison
Lucid
Sistemas Solicitados - Andamento - Documentos Google

Concursos UEM

UEM/PRH/DRH/Divisão de Recrutamento e Seleção
UEM - Concursos

LGPD

Publicações da ANPD — Autoridade Nacional de Proteção de Dados
2021.05.27GuiaAgentesdeTratamento_Final.pdf
LGPD - Google Drive

Notícias NPD

Núcleo de Processamento de Dados lança três novos Sistemas de Informação

MySQL

localhost / localhost | phpMyAdmin 4.9.5deb2

Perícia

Fui nomeado Perito, como proceder ? | Jusbrasil
iLovePDF

PROFURG

CPR - Login

REDCap

Partners – REDCap
REDCap Community
REDCap
RedCap-UEM-phpMyAdmin
REDCap Community

ZK

http://localhost:8084/
Documentation ZK

Sistemas Novos - Planilhas Google
Central de Estágio

HCK

Login · Técnicas de Invasão
(341) Vencendo um Desafio Hacker Juntos ao Vivo | TDI Fundamentos - YouTube
HackerOne | #1 Trusted Security Platform and Hacker Program
Have I Been Pwned: Check if your email has been compromised in a data breach
Shodan Search Engine
JSON Formatter & Validator
JSON Formatter & Validator
Forum TDI
Advangle: advanced web-search in Google and Bing
Who Is Hosting This? - Digital.com
Who Is Hosting This? - Digital.com
Whois scrap.net
Forum Técnicas de Invasão 🎩
Calça Color Five Pockets Classica - pierrecardin
0. Técnicas de Invasão

Reforma 2023

Revestimento-MATONE TIJOLOS - PIASTRA - Google Shopping
Revestimento-MATONE TIJOLOS - PIASTRA - Pesquisa Google
Tijolinho Vermelho De Barro - Revestimento Rústico Caixa 1,11m² | Leroy Merlin
Revestimento Bricks Natural Tijolinho Tijolo 1m² Plaquetinha
Revest Brick
Matone Tijolos - Construção & Decoração

CAPES

1.Dados Abertos
2. Avaliação da Pós-Graduação Stricto Sensu
https://dadosabertos.capes.gov.br/api/1/util/snippet/api_info.html?resource_id=920dd71a-79ed-47bb-9ba1-cb4b8277d2f5

Cursos

PHP-B7Web
Power BI-Udemy
Qlik Sense-Yoshida-Hotmart
Udemy

Geo

01. Extração de dados

01.Suicide Rates by State
02. GDP
03. Unemployment
97. U.S. Unemployment - Statistics & Facts | Statista
98. U.S.: Unemployment rate by age | Statista
99. Unempl
This Map Shows United States Gun Ownership by State [2020 Stats]
WISQARS Fatal Injury Reports-Suicide-FIrearm
Sub-national HDI - Subnational HDI - Table - Global Data Lab
WISQARS Fatal Injury Reports
Firearms Trace Data - 2012 | Bureau of Alcohol, Tobacco, Firearms and Explosives
Firearms Trace Data - 2020 | Bureau of Alcohol, Tobacco, Firearms and Explosives
rural - Census Bureau Tables

TOPODATA - Banco de Dados Geomorfométricos do Brasil
01.Aula Gratuita: Introdução aos Conceitos de Análise Espacial - YouTube
E-book: Análise Espacial de Dados Geográficos – ClickGeo: Cursos e Treinamentos em Produção de Mapas e Geoprocessamento
E-book: Introdução à Ciência da Geoinformação | ClickGeo: Cursos e Treinamentos em Produção de Mapas e Geoprocessamento
Livro Online: Guia Completo sobre Análise Geoespacial | ClickGeo: Cursos e Treinamentos em Produção de Mapas e Geoprocessamento
Introdução à Análise Multicritério em Geoprocessamento | ClickGeo: Cursos e Treinamentos em Produção de Mapas e Geoprocessamento
30-Meter SRTM Elevation Data Downloader
Earthdata Search
(1312) Como Baixar o NASADEM - YouTube
11: Livro: Análise Espacial com o programa R - MundoGEO

Chapter 5 Geometry operations | Geocomputation with R
01. Geocomputation with R
02. Geocomputation with R - 2022
A Guide to Working with US Census Data in R
Accessing and Analyzing US Census Data in R
Key Signup
Mapping Census Bureau Data in R with Choroplethr

Choropleth

13.18 ÚTIL - Creating a Choropleth Map | R Graphics Cookbook, 2nd edition
trulia/choroplethr: vignettes/i-creating-your-own-maps.Rmd
RPubs - Choropleth Vignette Dev
Choroplethr – AriLamstein.com
The choroplethr package for R - Trulia's Blog

GeoBR

Introduction to geobr • geobr

Geocomputarion with R

Chapter 12 Statistical learning | Geocomputation with R
Data Access - Vital Statistics Online
3D Visualization of US Population Density Per County Using ggplot2 + rayshader [OC] : dataisbeautiful

Plot US map with ggplot2

2. Mapping the US
How to plot state-by-state data on a map of the U.S. in R - Storybench
3 ggplot2 | Making Maps with R
r - How do I use ggplot2 to draw a US Map that uses two colors to fill in specific states based on a yes or no variable - Stack Overflow
How to Create a Map using GGPlot2: The Best Reference - Datanovia
Introduction to Creating Maps with ggplot2
Chapter 41 Different Ways of Plotting U.S. Map in R | Community contributions for EDAV Fall 2019

3 cursos de Análise Espacial em R
CURSO:An Introduction to Spatial Data Analysis and Visualisation in R | CDRC Data
Como calcular taxa por 100 mil habitantes
Cálculo de Mortes por Milhão de Habitantes - YouTube
Curso-Practical 1
RPubs - Violent Crimes in the United States (2014)
ggplot2 - How to map all the states of US using R with the number of crimes occurred in each state? - Stack Overflow
[Solved] How to map all the states of US using R with the number of crimes occurred in each state? - Local Coder
r - Display values corresponding to the USA states over the state name - Stack Overflow

USA-CENSUS

Contact Us | U.S. Census Bureau
Census Help
2000 United States census - Wikipedia
Census Bureau Delivers Colorado's Census 2000 Population Totals for Legislative Redistricting; First Race and Hispanic Data - Census 2000 - Newsroom - U.S. Census Bureau
Census Data for the United States | Data Emporium
2022 World Population by Country
List of U.S. states and territories by population - Wikipedia
State Population Totals: 2010-2019
Historical Population Change Data (1910-2020)
Census Bureau Search

00. Mortality Multiple Cause Files

Suicide Rates by State
00. Multiple Cause of Death, 1999-2020 Request Form

01. Resident Population Tables

Resident Population Tables-2000

02.County Population Totals: 2010-2019

County Population Totals: 2010-2019

03.Single-Race Population Estimates 2010-2020 by State and Single-Year Age Request

Saved Request: Single-Race Population Estimates 2010-2020 by State and Single-Year Age, D279F915
Single-Race Population Estimates 2010-2020 by State and Single-Year Age Results Form

Dados em Painel

(130) Dados em Painel - Aula 08 - Aplicação no R - YouTube
Econometrics Academy - Simple Panel Data Models
Capítulo 6 Regressão com Dados em Painel | Software R: curso avançado

Artigos RQ

i3-1e941ade6f1aa8ea2da3a6a517b515df.pdf
Microsoft Word - Artigo_Revisado _1_.docx - 35004-Texto do Artigo-117390-1-10-20170404.pdf
WagnerFragaEconomia - WagnerFraga.pdf
477765948008.pdf
Submissao819.pdf
DEPARTAMENTO DE CIÊNCIAS ECONÔMICAS - Monografia Rafael Feltrin.pdf
ppp_n62_Artigo3-determinantes_socioeconomicos.pdf
93655-Texto do artigo-493391-1-10-20211029.pdf
07 Amaral_et_al_v13_n2_p288-308_2019 - inacio,+07+Amaral_et_al_v13_n2_p288-308_2019.pdf
inacio,+Manuscript_607.pdf
Carlos Ramalheira - Epidemiologia do Suicídio em Portugal.pdf
Analiseeconometricoespacial_Araujo_2020.pdf
404 Not Found

Software - Bayesian Intelligence
EmelynneGabriellyDeOliveiraSantos_DISSERT.pdf

MDP

Cias Áereas

Aeromexico
American Airlines-3004-5000 (do Brasil)
Air France Brasil
Azul - Linhas Aéreas Brasileiras. O Céu é Azul.
Copa Airlines
GOL
KLM Royal Dutch Airlines
LATAM-0300-570-5700 ou 4002-5700
LATAM
Lufthansa
Qatar Airways
SWISS
TAP - Tabela Dinâmica
Turkish Airlines ®️ | Voando para a Maioria do Países‎
United Airlines
TAP-Reservar com milhas

FlightConnections
Google Flights
Skyscanner
Gol - FlightConnections
Azul - FlightConnections
LATAM - FlightConnections
LATAM - Whats - 56-9-6825-0850
Decolar
MDP
123Milhas

Diversos

Earth
Brazil Weather Map & Radar
Para ver Taxa de Embarque
Call Portugal directly using Skype | Starts from R$ 0,06/min
SumUp | A máquina de cartão número 1
Resgate seus pontos e compre passagens aérea | LATAM Pass
CVC

Star Alliance – Conheça os produtos | TAP Air Portugal
Máquina De Passagens
Parceiras-LATAM Pass
Tabela Fixa LATAM
Parceiras-Gol
Tabela de Regiões | Viagem | LATAM Pass
AAdvantage – American Airlines

MDM

(1) Máquina de Milhas | Facebook
Hotmilhas
Maxmilhas
Livelo
Esfera
LATAM
LATAM Pass
LATAM-Tabela de Regiões
Smiles
As melhores ofertas de Cashback, Pontos e Milhas em um só lugar
Km de Vantagens
Máquina De Milhas
Perguntas Frequentes
TudoAzul
Interline TudoAzul
LATAM
Azul

Concurso RF

Cursos para Analista de TI da Receita Federal | Estratégia Concursos

APB

Saiba como descobrir se a empresa está em dia com o imposto de renda
QlikView
Certidão de Regularidade Fiscal — Receita Federal
Certidão Negativa de Débitos Trabalhistas - TST
CPOM PoA

DPOM - PoA

CPOM - Envio da Documentação - SMF - Portal de Serviço
SMF

DPOM - PoA
EBPP

Artigos

Enviado

Para Ler

Annotated - Legal and ethical concerns of complex predictive analytics - Health Affairs 2014.pdf - Google Drive

Lido

Revistas

Artificial Intelligence in Medicine - Journal - Elsevier
JAMIA
Elsevier Account Settings
Knowledge-Based Systems

Publicado

Standardizing clinical trials workflow representation in UML for international site comparison
Application Description and Policy Model in Collaborative Environment for Sharing of Information on Epidemiological and Clinical Research Data Sets
Standardizing Clinical Trials Workflow Representation in UML for International Site Comparison
de Carvalho: CURSO DE ATUALIZAÇÃO EM METODOLOGIA... - Google Scholar
Reenginering Clinical Research Teams: An Organizational Modeling Approach - IEEE Conference Publication
Sistema de Información y Educación Superior: análisis de datos y caracterización de la “Universidade Estadual de Maringá”
Carvalho: Modelagem de um Cluster Heterogêneo para... - Google Scholar
Particionamento de grafos de aplicações e mapeamento em grafos de arquiteturas heterogêneas
Regional Disparities in Mortality after Ischemic Heart Disease in a Brazilian State from 2006 to 2010
Workflow in Clinical Trial Sites & Its Association with Near Miss Events for Data Quality: Ethnographic, Workflow & Systems Simulation
VARIABLES INVOLVED IN THE MANAGEMENT OF SCHOOL BULLYING: A BAYESIAN NETWORK ANALYSIS
Track Your Article
Brazilian Journal of Psychiatry
Journals Article Sharing | Policies | Elsevier
Produção
BNPA
Knowledge-Based Systems | Vol 223, 8 July 2021 | ScienceDirect.com by Elsevier

Qualis

Plataforma Sucupira

Para Publicar

ELSEVIER Knowledge Systems

Guide for authors - Knowledge-Based Systems - ISSN 0950-7051
Database Linking
Data Visualization | Elsevier
Illustration Services - Webshop | Elsevier
Artwork and media instructions
Highlights
Submit your paper
Plagiarism detection

SEM

An overview of structural equation modeling: its beginnings, historical development, usefulness and controversies in the social sciences | SpringerLink
REVISAO SISTEMÁTICA - Causalidade e Fonoaudiologia: abordagem epidemiológica

Avaliar

RPubs - Multiple Linear Regression on Auto data
Exemplo de Regressão Linear Simples
(PDF) EXERCICIOS CORRELAÇÃO E REGRESSÃO LINEAR | Erick Neves - Academia.edu
6. Exercícios - Análise de Regressão | Portal Action
(PPT) A Selecao do Teste estatistico | Silvania Goularte Correia - Academia.edu
Qual teste estatístico devo usar? – Sobrevivendo na Ciência
notaR
7. Modelos Lineares [Using R]
Revealing Facts and Avoiding Biases: A Review of Several Common Problems in Statistical Analyses of Epidemiological Data
An overview of robust methods in medical research. - PubMed - NCBI
Causal Models and Learning from Data
Princípios de Estatística em Ecologia - 9788536324326 - Livros na Amazon Brasil
GitHub - renanxcortes/springerQuarantineBooksR: Download all books made available by Springer during the COVID-19 quarantine!

LER - 15 - 05 - 2020

Intersections of machine learning and epidemiological methods for health services research | International Journal of Epidemiology | Oxford Academic
Causal inference methods and simulation approaches in observational health research within a geographical framework - White Rose eTheses Online
A Statistical Roadmap for Journey from Real-World Data to Real-World Evidence | SpringerLink
Causality and Speech, Language and Hearing Sciences: epidemiological approach
Real-World Evidence, Causal Inference, and Machine Learning - ScienceDirect
Epidemiologic Toolbox: Identifying, Honing, and Using the Right Tools for the Job | American Journal of Epidemiology | Oxford Academic
Counterfactual Reasoning for Process Optimization Using Structural Causal Models | SpringerLink

Bioestatística descritiva | DENDRIX
(63) INTERNET LENTA? TROQUE O CANAL DO SEU ROTEADOR E MELHORE SEU SINAL WI-FI - YouTube

BC

Bradesco
Banco Bradesco S/A
Itaú
Caixa Econômica
Nubank
Santander
Bradesco Internet Banking

Comida

Macarrão de forno com hambúrguer • Cozinha Legal
Molho Branco
Nhoque de batata com molho branco • Cozinha Legal
Macarrão com presunto, queijo e ervilhas • Cozinha Legal
Macarrão de forno com molho bolonhesa e creme de queijo • Cozinha Legal
Macarrão de forno cremoso com presunto e queijo • Cozinha Legal

WhatsApp

Cursos

Curso de Machine Learning Online (com R)
Curso-R | Top 10 pacotes para data science

Hostgator

EAD-EAE-MGA

Moodle
EAD-Arautos-Maringá: Administração: Pesquisa
Webmail - EAD
Moodle Admin
(1) Como criar curso no Moodle | EnsineOnline - YouTube
pAPEL DOS Participantes

Moodle

01. Adm - Usuários
02. Moodle Participantes
Curso de Moodle - Criando Plataforma de Cursos - YouTube
Manual do Professor - Moodle: Inserir usuários
(1) Tutorial Moodle: Ativando o modo de Edição - YouTube
(1) Hugo Vasconcelos - YouTube
(1) Curso de Moodle - Personalizando o Tema - YouTube
(1) Como configurar o OBS para gravar ou transmitir Powerpoint e Webcam - YouTube
Entenda como utilizar o Moodle - Dante Explica - YouTube
Temas Plugins
CURSO MOODLE #5 - CONFIGURAÇÃO DE PERFIL DE USUÁRIO - YouTube
(1) Atividades no Moodle - YouTube

Softaculous - Softaculous - Moodle
Portal do Cliente

Solicitar Certificado SSL - HostGator Brasil
Como solicitar o certificado SSL grátis? – Suporte HostGator

Escola de Pesquisa Cientifica

EPC - WordPress
100+ Online Business Pictures | Download Free Images on Unsplash

(25) Como Criar Um Site Portfólio Para Os Seus Clientes (HostGator) - YouTube

IT4BSR

IT4BSR – Information Technology for Business and Scientific Research
Webmail
WordPress
(9) Como Criar um Site Completo (ELEMENTOR) - YouTube

Instagram

Perfil comercial Instagram: Entenda por que e como aderir agora (com Passo a Passo) - Leadlovers
Estúdio de Criação do Facebook
PPGSS/UERN (@ppgss.mestrado.uern) • Fotos e vídeos do Instagram
Mestrado em Atenção à Saúde (@mestradosaudepuc) • Fotos e vídeos do Instagram
Mestrado em Saude Pública (@mestradosaudepublicaunioeste) • Fotos e vídeos do Instagram
Mestrado Ciências da Saúde (@mestradosaudeunoeste) • Fotos e vídeos do Instagram
MESTRADO EM NUTRIÇÃO E SAUDE (@ppgns_ufes) • Fotos e vídeos do Instagram
MESTRADO EM SAÚDE COLETIVA EAD (@mestrado.saude.ead) • Fotos e vídeos do Instagram
Mestrado em Ciências da Saúde (@cienciasdasaudepucc) • Fotos e vídeos do Instagram
Mestrado em Genética - MGene (@mestradoemgenetica.mgene) • Fotos e vídeos do Instagram
Mestrado em Enfermagem UDESC (@ppgenfaps) • Fotos e vídeos do Instagram
Mestrado em Enfermagem UFMA (@ppgenfufma) • Fotos e vídeos do Instagram
Mestrado em Psicologia UFGD (@ppgpsico.ufgd) • Fotos e vídeos do Instagram
Mestrado GESTEC/UNEB (@mestradogestecuneb) • Fotos e vídeos do Instagram
Mestrado - UNIOESTE (@mestrado.foz.unioeste) • Fotos e vídeos do Instagram
Mestrado em Odontologia UFPI (@ppgoufpi) • Fotos e vídeos do Instagram
Mestrado QualiSaúde (@mestradoqualisaude) • Fotos e vídeos do Instagram
Ca
Home - Canva

Facebook

IT4BSR-Information Technology for Business and Scientific Research | Facebook

Free-Pictures
59 melhores bancos de imagens gratuitos de 2020!
WPTest
Projeto Dupla
Sobre ‹ Faça DIREITO — WordPress

INSS

MEU INSS

IRPF

UNIMED
2021-Consulta Restituição
Extrato IRPF
Consulta Restituição

PROFURG

FAAD-Moodle

Admin Site
Plone Community
Poster - In case of emergency
FAAD
Flyer - PROFURG Flyers

AED

Análise exploratória de dados
4. Análise Exploratória de Dados [Using R]
A protocol for data exploration to avoid common statistical problems - Zuur - 2010 - Methods in Ecology and Evolution - Wiley Online Library
4. Tutoriais de Análise Exploratória de Dados [Using R]
Análise Exploratória de Dados no R
Exercícios: estatística básica
Estatística Básica - Bussab e Morettin (2017)
2 Análise descritiva
(104) Which graphical images should I use to show my results? [Portuguese] - YouTube
Guia de Bolso: ggplot2 | Gráficos elegantes no R
How to Order the Bars in a ggplot2 Bar Chart - Statology
Data Visualization With R - Bar Plots - Rsquared Academy Blog - Explore Discover Learn
Preparing plots in R for statistical reports: a tutorial
R tips: 16 HOWTO’s with examples for data analysts
(36) GRÁFICO BOXPLOT ✅ EXERCÍCIO RESOLVIDO - YouTube
(36) Gráfico de Box Plot - YouTube
Machine Learning com dados da base Titanic | Kaggle

Curso Anova

Estatística básica no R – Curso gratuíto – A N O V A
Psicometristas Brasil
(1263) 37 - Tamanho do efeito - d de cohen - YouTube
(1263) tamanho do efeito anova - YouTube
(1263) ANOVA de duas vias no R (Parte 2) - YouTube
(1263) Canal Pesquise - YouTube
(1263) Análise de Variância MOLEZA (JAMOVI) - YouTube
Cap. 13 ANOVA de medidas repetidas | Conceitos e análises estatísticas com R e JASP
How to Perform a Repeated Measures ANOVA in R - Statology
(1263) ANOVA de uma via com medidas repetidas no R - YouTube
Cap. 13 ANOVA de medidas repetidas | Conceitos e análises estatísticas com R e JASP
Cap. 12 ANOVA | Conceitos e análises estatísticas com R e JASP
(1577) Análise de Variância MOLEZA (JAMOVI) - YouTube
(1577) ANOVA de medidas repetidas fácil como você nunca viu! (JAMOVI) - YouTube

Mann-Whiney

2 - Teste de Wilcoxon-Mann-Whitney - Amostras Independentes - Técnicas Não Paramétricas | Portal Action

Paula Abrante

rfsaldanha-microdatasus
rfsaldanha-microdatasus-Manual de utilização do pacote
Política Nacional de Humanização - HumanizaSUS — Português (Brasil)

Registro de domínio - Compre seu Domínio à partir de R$ 6,99

Nova pasta

Statistics for Nursing Research - E-Book: A Workbook for Evidence-Based Practice - Susan K. Grove, Daisha J. Cipher - Google Books
Statistics for the Behavioral Sciences - Frederick J Gravetter, Larry B. Wallnau - Google Books

Tamanho do Efeito

Tamanho do Efeito (Effect Size) - tamanho_do_efeito.pdf
RPICS_vol3_1_2016_5_TDE - 151538734.pdf
(PDF) Calcular e apresentar tamanhos do efeito em trabalhos científicos (3): Guia para reportar os tamanhos do efeito para análises de regressão e ANOVAs | Revista Portuguesa de Investigação Comportamental e Social / Portuguese Journal of Behavioral and Social Research - Academia.edu
Effect Size
Como interpretar e relatar eta ao quadrado / eta ao quadrado parcial em análises estatisticamente significativas e não significativas?
(1263) Estatística Básica| TAMANHO DO EFEITO: TESTE T, CORRELAÇÃO, ANOVA E QUI-QUADRADO - YouTube
anova - How to interpret and report eta squared / partial eta squared in statistically significant and non-significant analyses? - Cross Validated
Effect Size in Statistics - The Ultimate Guide
FAQ/effectSize - CBU statistics Wiki
Frontiers | Calculating and reporting effect sizes to facilitate cumulative science: a practical primer for t-tests and ANOVAs | Psychology
(1263) Eta Squared Effect Size for One-Way ANOVA (12-7) - YouTube
effectsize.pdf
Calculating Eta Squared in R for repeated measures - Pesquisa Google
How to Calculate Eta Squared in R - Statology
EtaSq: Effect Size Calculations for ANOVAs in DescTools: Tools for Descriptive Statistics
(1263) Tamanhos de efeito para o teste de Mann-Whitney no R - YouTube
etaSquared function - RDocumentation
(1263) R - eta squared (as effect size for one-way anova) - YouTube
(22) (PDF) Calcular e apresentar tamanhos do efeito em trabalhos científicos (3): Guia para reportar os tamanhos do efeito para análises de regressão e ANOVAs [Calculating and reporting effect sizes on scientific papers (3): Guide to report regression models and ANOVA effect sizes]
Effect Size Statistics for Anova Tables #rstats | R-bloggers
Effect sizes for ANOVAs

REDCap

Recorded Lectures – REDCapCon Virtual
PMA Local
NginX allow POST on static pages. · GitHub
REDCap
Pesquisa - 295 Resultados encontrados - REDCap
REDCap Community Site – REDCap
REDCap
Entrar - REDCap Community
REDCap-Mandaqui

Trello

UEM

PDI-Antigo
PDI-Antigo-PHPMyAdmin
RedMine
TELEFONES
Db2 SELECT DISTINCT: Prevent Duplicate Rows Returned by a Query
PRH

ABNT

Referencias bibliograficas - exemplos (livros, e-books, blogs, internet)

ANOVA

(139) ANOVA de uma via no R - YouTube

Apoio Artigo Knowledge Based Systems

[PDF] Using Directed Acyclic Graphs for Investigating Causal Paths for Cardiovascular Disease | Semantic Scholar

Elementor

(35) Como Criar um Blog do Zero - Tutorial Passo a Passo - YouTube
Curso de Elementor Pro - Crie Sites Incríveis! - a Escola de Sites
(39) Criando o Seu Primeiro Site - Passo a Passo (Módulo 1) - YouTube
Elementor Pro - A Escola De Sites

Mestrado Profissional

Corpo Discente – UFPR – Programa de Pós-Graduação em Saúde Coletiva
edital-ppgsc-18-2019.pdf

MCA

The GoodmanKruskal package: Measuring association between categorical variables
(107) Latent Class Analysis in R with poLCA package for beginners - Part 2 - YouTube
Int_er.net....ban:ki_ng___cAI:XA

Mint19

6 Dicas [Pouco utilizadas] para melhorar sua internet no Ubuntu LINUX
Fórum Linux Mint Brasil

Qlik

Qlik Management Console

Qlik Sense Hub
PDI 2019 · Painel · Metabase

Dicas e Truques Visuais - Qknow
Operadores relacionais ‒ QlikView
University of East London
Dashboard - Curso Qlik Sense
Course: Business Analyst Library

Treinamento

(99) Qlik Sense - Conceitos Básicos - YouTube
Hotmart Club - Aviso 2 sobre Expiração Qlik Sense Desktop

Qlik Sense | Boas práticas para melhorar a experiência do usuário - BIX Tecnologia

Apps

PDI

https://npd.uem.br/pdi/indexUnidades.zul
(72) Qlik Sense – Product Tour - YouTube
Qlik Sense Hub - DEMO
8 dicas para deixar seus dashboards eficientes - Cluster - Qlik Sense advanced web dashboards
Qlik Sense do Básico ao Avançado com Extras de Especialista | Udemy
LOGOTIPOS
SenseTheme
Video Qlik Sense! - YouTube

Videos

(136) Dominando o Set Analysis no Qlik Sense - YouTube

Vagas

2 vagas para Qlikview/Qlik Sense até 12mil - ABRACD - ASSOCIAÇÃO BRASILEIRA DE CIÊNCIA DE DADOS

Qlik Sense: quais são as certificações disponíveis?
inQuesti | Qlik Sense - Dashboards, relatórios e self-service BI
Qlikview 11 for Developers.pdf
Mastering Qlikview | Amazon.com.br
Qlik Sense Cookbook (English Edition) - eBooks em Inglês na Amazon.com.br
Atualização_Qlik_Sense
Parceiros - BI com Vatapá
Mastering Qlik Sense | Amazon.com.br
Obtendo todas as Certificações da Qlik
Onde Estudo - BI / DW / Big Data - BI com Vatapá
Qlik Sense & Set Analysis: O Guia Definitivo eBook: de Freitas, Andrey Rodrigues: Amazon.com.br: Loja Kindle
Qlik Sense Hub
(48) Total Acumulado por Mês de Cada Ano em Qlik Sense - YouTube
Assistir The Walking Dead: World Beyond 1x1 Online – Cine Vision
SenseTheme
Installing custom themes - Qlik Sense - YouTube
Customizar o Thema para o Qlik Sense - YouTube
Maq. Dorival
SenseTheme

Stat Sem Mat

SAGE - Statistics for the Health Sciences

Stat

Comparando dois grupos
RPubs - Exemplo de regressão linear múltipla

Simple Linear Regression

Como fazer uma análise de regressão linear simples no R
Tutorial — Ajuste e Interpretação de Regressão Linear com R

RPubs

Logistic Regression

RPubs - Regressão Logística aplicada ao caso da Diabetes

Regressão Linear Simples

7a. Regressão Linear Simples [Using R]

Regressão Linear Múltipla

O que é o teste F de significância global na análise de regressão?
Teste F na Regressao Linear
Microsoft Word - custos_300.doc - Utilização da regressão linear como ferramenta de decisão na gestão de custos.pdf
Aula 07 - Modelando
Explaining the lm() Summary in R – Learn by Marketing

Comparação de Grupos

R Handbook: One-sample Wilcoxon Signed-rank Test

Cursos

EcoR

Poder Estatístico

De quantos dados preciso no meu projeto? – Sobrevivendo na Ciência
Tamanho do Efeito

Livros

books
SAGE - Student and Instructor Site for Statistics for the Health Sciences

Test-T

Teste t de Student no R
RPubs - Mini tutorial de test T no R

Tamanho do Efeito

Computation of different effect sizes like d, f, r and transformation of different effect sizes: Psychometrica
hcistats:wilcoxonsigned [Koji Yatani's Course Webpage]
https://www.youtube.com/watch?v=rEEFxfgTdRg
https://www.youtube.com/watch?v=39qt9fF4ThU
https://www.youtube.com/watch?v=1zbpiNY2YWE
Test and effect size details
Effect size for Wilconxon signed-rank test — Forum
Wilcoxon Effect Size — wilcox_effsize • rstatix

P-Valor

Como Interpretar a Significância Estatística

Missing Data

Como lidar com valores faltantes (missing)? – Portal Data Science
Multiple data imputation and explainability | R-bloggers
Handling missing data with MICE package; a simple approach | R-bloggers
Imputing missing data with R; MICE package | R-bloggers

UCI Machine Learning Repository

Desbalanceamento de Classes

01. Learning from Imbalanced Classes - Silicon Valley Data Science
02. 11 Subsampling For Class Imbalances | The caret Package
machine learning - Opinions about Oversampling in general, and the SMOTE algorithm in particular - Cross Validated
The Precision-Recall Plot Is More Informative than the ROC Plot When Evaluating Binary Classifiers on Imbalanced Datasets
Minerando Dados › Lidando com Classes Desbalanceadas – Machine Learning
Classificação de Deputados Eleitos - 2014 utilizando caret package
normal distribution - Linear regression with unbalanced dummy variables + not normally distributed residuals - Cross Validated

Análise Exploratória de Dados

(27) Tutorial R | Lidando com dados faltantes Visdat no R - YouTube
visdat package | R Documentation
Using visdat

Random Forest

(1) Uma aplicação com Random Forest no R - YouTube
(1) EXCLUSIVO: PADRE ARAUTO RELATA EXORCISMOS - YouTube
(1) R Tutorial. Random Forest Regression - YouTube
Entenda como funciona o Random Forest (Machine Learning)
(2) Árvore de decisão - Como funciona (Machine Learning) - YouTube
Editing streams ‒ Qlik Sense on Windows
Sensitivity vs Specificity and Predictive Value - Statistics How To
00.Tune Machine Learning Algorithms in R (random forest case study)
Model tuning: Otimizando os hiper parâmetros do modelo – hacking analytics
Aprendendo em uma Floresta Aleatória - Machina Sapiens - Medium
How to implement Random Forests in R | R-bloggers
Bagging and Random Forest Ensemble Algorithms for Machine Learning
Tutorial: Como Fazer o Seu Primeiro Projeto de Data Science
new approach for interpreting Random Forest models and its application to the biology of ageing | Bioinformatics | Oxford Academic
StatQuest: Random Forests Part 1 - Building, Using and Evaluating - YouTube
A small guide to Random Forest - part 1 - PaolaElefante.com
A small guide to Random Forest - part 2 - PaolaElefante.com
Tree Based Algorithms : A Complete Tutorial from Scratch (in R & Python)
(4) Classificação com Random Forest - R - YouTube
(4) Random Forest no R - YouTube
(4) Tutorial R | Machine learning no R - YouTube
Aula 10 - Data Science - R - Caso do Titanic - Kaggle - Código Fluente
(7) Parte 4 - Árvore de Decisão, Random Forest e Gradient Boosting - Agora ou nunca! - Prof. Dri - YouTube
Aula 09 - Data Science - R - Caso do Titanic - Kaggle - Código Fluente
00.Random Forest - Construindo e Avaliando - YouTube
r - Caret and randomForest number of trees - Cross Validated
Implementing a Random Forest Model
7 train Models By Tag | The caret Package
R : Train Random Forest with Caret Package (R)
Using Caret & RandomForest in R for (top 10%) | Kaggle
random forest caret package - Pesquisa Google
Navigating The Random Forest Algorithm in R - Towards Data Science
RPubs - Análise de Dados 2 - Problema 5
Internet---BAnking....CaIXA
A complete guide to Random Forest in R
MedCalc's Diagnostic test evaluation calculator

Razao de Chances

Guia para iniciantes sobre como interpretar razão de chances, intervalo de confiança e valor p. - Students 4 Best Evidence

POLCA

latent class analysis with the poLCA-package in R - Pesquisa Google
Example for a latent class analysis with the poLCA-package in R – ahoi data
Latent Class Analysis Frequently Asked Questions (FAQ)
Latent Class Analysis Frequently Asked Questions (FAQ)
(PDF) randomLCA : An R Package for Latent Class with Random Effects Analysis
Application of latent class analysis in assessing the awareness, attitude, practice and satisfaction of paediatricians on sleep disorder management in children in Italy
(PDF) Latent Class Analysis: An example for reporting results

Nova pasta

electrolux dm94x - Pesquisa Google

Importado do Internet Explorer

Galeria do Web Slice
Sites Sugeridos

Diversos

Start | R Psychologist
R Datasets
Consultant - Business Intelligence - Basking Ridge, NJ 07920 - Indeed.com
Qlik Administration Lead - Hoboken, NJ 07030 - Indeed.com
BI Developer - Eglin AFB, FL 32542 - Indeed.com
Developer, Business Intelligence - Dover, DE 19904 - Indeed.com
Qlik Sense and Tableau Developer - Remote - Indeed.com
Senior Data Analyst, BI - Bellevue, WA 98008 - Indeed.com
Technical Data Analyst - Overland Park, KS - Indeed.com
Business Intelligence Analyst - Encino, CA 91436 - Indeed.com
Business Intelligence Analyst - Rolling Meadows, IL 60008 - Indeed.com
BI Developer - Sayre, PA - Indeed.com
Business Intelligence Specialist - Chicago, IL 60654 - Indeed.com
QLIK/BI ANALYST III - Greeley, CO 80634 - Indeed.com
Business Intelligence Developer - Norfolk, VA - Indeed.com
Business Intelligence Engineer - New York, NY - Indeed.com
QlikSense Developer | Bridgewater, NJ - Bridgewater, NJ - Indeed.com
Page Not Found - Indeed Mobile
Associate Business Intelligence Specialist (Tableau, Qlik, SAS, Cognos) - Glen Allen, VA 23060 - Indeed.com
QlikSense Developer | Bridgewater, NJ - Bridgewater, NJ - Indeed.com
BI Qlik Sense Developer - Arlington, VA - Indeed.com
Qlik Developer - Saint Petersburg, FL 33716 - Indeed.com
Qlik Sense Developer - Chicago, IL 60606 - Indeed.com
Qlik Sense Developer - Buffalo, NY - Buffalo, NY 14201 - Indeed.com
Qlik Sense Engineer/Administrator (Remote) - United States - Indeed.com
Qlik Sense Developer - Huntsville, AL 35808 - Indeed.com
Qlik Sense Developer - McLean, VA 22102 - Indeed.com
Qlik Sense Developer - Remote - Indeed.com
Qlik Sense Developer - McKinney, TX - Indeed.com
Qlik Sense / Reports Developer - Seaside, CA - Indeed.com
Qlik Sense Jobs, Employment | Indeed.com
"Qlik View/Qlik Sense Developer" "Remote Job" - Pesquisa Google

Tópico 6 Editar nome do tópico
Nubank_2021-10-02.pdf
Big Box USA | Sua opção para redirecionamento de encomendas.
TWD11
Messages
Blog de Geografia: Jogos Geográficos
Sugestões de Avaliação - Geografia - 7º ano - Projeto Araribá - MODELO 1
geo_ano7_uni1.indd - geo_ano7_uni1.pdf
geo_ano6_uni8.indd - geo_ano6_uni8.pdf
Ara_geo_6_uni_7_edit_diag.indd - ara_geo_6_uni_7_edit_diag.pdf
geo_ano6_uni6.indd - geo_ano6_uni6.pdf
Ara_geo_6_uni_5_edit.indd - ara_geo_6_uni_5_edit.pdf
Arariba_Geo_6_unid 4_edit.indd - arariba_geo_6_unid_4_edit1.pdf
geo_ano6_uni3.indd - geo_ano6_uni3.pdf
(2151) Jogos Didáticos de Geografia: Conheça! Experimente! Inove! - YouTube
electrolux dm84x - Pesquisa Google
Sign in to your Qlik account
Como instalar e configurar o MariaDB Server no Linux (MySQL) - YouTube
Cebraspe
Como pagar uma fatura – Suporte HostGator
Seja um investidor da Kuba
SuperFlix Oficial © - Filmes e Séries Online em um só lugar. - Assistir filmes e séries online aqui é diversão garantida, o © SuperFlix Oficial é referência em filmes online em Full HD no Brasil. Filmes Lançamentos 2022.
Oi Modem
Speedtest by Ookla - The Global Broadband Speed Test
Teste sua Velocidade de Internet | SIMET
Internet lenta? A gente te ajuda! | ConectaJá - Tudo que importa sobre sua internet e telefonia
As melhores ofertas de Cashback, Pontos e Milhas em um só lugar
WhatsApp Web

INÍCIO - Sem Fronteira
Philips X-treme Vision H13
DMS
Fotos-UEM

R

R for data science: getting started, EDA, data wrangling - Library Calendar - Duke University
R for data science: visualization, pivot, join, regression - Library Calendar - Duke University
Rfun
Rfun

Sub100

Casa à venda em Jardim Paulista Maringá - PR | Sub100 Imóveis
Sobrado à venda em Vila Esperança Maringá - PR | Sub100 Imóveis
Casa à venda em Jardim Santa Helena Maringá - PR | Sub100 Imóveis
Casa à venda em Jardim Colina Verde Maringá - PR | Sub100 Imóveis
Casa à venda em Jardim Paris Maringá - PR | Sub100 Imóveis
Casa à venda em Jardim Santa Helena Maringá - PR | Sub100 Imóveis
Casa à venda em Jardim São Clemente Maringá - PR | Sub100 Imóveis
Casa à venda em Jardim Liberdade Maringá - PR | Sub100 Imóveis
Casa à venda em Jardim Paris VI Maringá - PR | Sub100 Imóveis
Casa à venda em Jardim Munique Maringá - PR | Sub100 Imóveis
Casa à venda em Jardim Paris Maringá - PR | Sub100 Imóveis
Casa à venda em Jardim Itaipu Maringá - PR | Sub100 Imóveis
Casa à venda em Jardim Munique Maringá - PR | Sub100 Imóveis
Casa à venda em Jardim Higienópolis Maringá - PR | Sub100 Imóveis
Casa à venda em Jardim Dias I Maringá - PR | Sub100 Imóveis
Casa à venda em Jardim Pilar Maringá - PR | Sub100 Imóveis
Casa à venda em Jardim Everest Maringá - PR | Sub100 Imóveis
Casa à venda em Jardim Espanha Maringá - PR | Sub100 Imóveis
Casa à venda em Jardim Olímpico Maringá - PR | Sub100 Imóveis
Casa à venda em Jardim Munique Maringá - PR | Sub100 Imóveis
Casa à venda em Jardim Munique Maringá - PR | Sub100 Imóveis
Casa à venda em Jardim Alvorada Maringá - PR | Sub100 Imóveis
Casa à venda em Jardim Munique Maringá - PR | Sub100 Imóveis
Casa à venda em Jardim Munique Maringá - PR | Sub100 Imóveis
Casa à venda em Jardim Oriental Maringá - PR | Sub100 Imóveis
Casa à venda em Parque Industrial 200 Maringá - PR | Sub100 Imóveis
Casa à venda em Loteamento Sumaré Maringá - PR | Sub100 Imóveis

Silvio Iwata

Sobrado à venda em Maringá - PR , Jd Imperial I
Apartamento à venda em Maringá - PR , Vl Bosque
Apartamento à venda em Maringá - PR , Zona 07
Apartamento à venda em Maringá - PR , Zona 02
Apartamento à venda em Maringá - PR , Zona 03
Imóveis residenciais à venda em Maringá - PR

Chapa 4 - Trabalho e Renovação | Podcast no Spotify

CN
PDI • Início
CEHU
STHDA
Compra de Pontos
pwr

RC

REDCap-Hostgator-Elias
REDCap - IC - FUC
REDCap - UEM
PDI (PLD)-EM REV - Plano Diretor Institucional - 9. Objetivos e Ações - Dados Analíticos | Pasta - Qlik Sense
REDCap Local
phpinfo()
PMA-IC-FUC
Gmail
IP Location Finder - Geolocation
IC-FUC-Novo
IC-FUC-phpMyAdmin-Novo
REDCap-Novo

RC-F

Financeiro - HostGator
REDCapHG
REDCap
phpMyAdmin
host server
Info.php
ProFTPD no Centos, RHEL e seus derivados - veja como instalar

Firewall

Firewall Básico no Redat7/CentOS7 – Cleberson Blog
How to set up a firewall using FirewallD on CentOS 8 - nixCraft

How to Step by Step Install MySQL 8 on CentOS 7
00.How To Install phpMyAdmin With Apache on CentOS 8 / RHEL 8 | ITzGeek
PhpMyAdmin Error The json extension is missing. Please check your PHP configuration. [Perguntas não tão freqüentes]
Como fazer uma configuração básica de Postfix no CentOS – Ricardo Martins
Postfix - Ativando porta 587 SMTP [Dica]
Forbidden :You don’t have permission to access /phpmyadmin on this server – oraerr.com
Sendmail won't send emails on CentOS 7 (Permission denied)
SELINUX: Set enforcing and permissive modes for SELinux. - CertDepot

IPTables

Firewall iptables no Linux - Parte 02: Criar e Visualizar uma regra básica - Bóson Treinamentos em Tecnologia

graficos aciclicos dirigidos - Pesquisa Google

Canal SEM
SOGo
Resultado do Vale Sorte deste domingo – 12/07/2020 - Mais Nacional
WP

Baixar video do Facebook online | Facebook video downloader
Converter MP4 em WEBM: conversor de MP4 para WEBM online gratis
BRQ – Carreiras
PicPay
Instruções por voz gratuitas, informações de trânsito e App de Navegação por GPS feito pelo Waze.
Domingo Espetacular – Entretenimento e Informação – Rede Record
Aula01.ipynb - Colaboratory
Inicial - Doctor's First
Painel
UEL Mobile – Apps no Google Play
INEDAE
Meet: zor-pjgr-fae
Project | Project Company
Holerites/Inf. Rendimentos | Meira Fernandes Gestão e Solução
Universidade Estadual de Londrina
Linux Mint 20.1 Xfce Release Notes - Linux Mint
New features in Linux Mint 20.1 Xfce - Linux Mint
Episódios Completos de The Walking Dead | Assista a temporada 10 Online - FOXPlay.com
Acesso.Gov.Br
(804) Dica 17 - RESUMO no LATTES: ELABORE UM DE QUALIDADE! - YouTube
REDCapCon
(871) Como restaurar boot do dual boot pelo grub quando ele desparece - YouTube
EVE Online | The #1 Free Space MMORPG | Play here now!
Assistir The Walking Dead: 10x16 Dublado e Legendado - Max Séries
The Walking Dead BRASIL - Tudo Sobre o Universo TWD!
Você pesquisou por THE WALKING DEAD - Max Séries
Tarefa
Royalty Free Stock Footage - Artgrid.io
Grego Santos. | Advogados em Maringá, São Paulo e Brasília
Arroz cremoso fácil | CLAUDIA
(56) A Ciência Por Trás - YouTube
New Job Position - ERA Chair | Faculdade de Medicina da Universidade de Lisboa
Qlik Sense Hub
QR Code Generator | Crie aqui QR Codes gratuitos
Ecai 2020 – Congreso Ecai 2020
Elo Meus Cartões
(933) Ana Vidović - FULL CONCERT - Live from St. Mark's - YouTube
Livresat
Oxford - Loja Oficial - Tudo para o seu dia a dia ser especial
"qlik sense" developer remote jobs - Pesquisa Google
Processo Seletivo | Apresentação em Vídeo

Introdução
Resgate seus pontos e compre passagens aérea | LATAM Pass
Aprofundando as técnicas sobre gráficos – Phylos.net

Fluency TV
HUM-CPR
Controle de Ponto
AssElGB

Hacker

Replit: the collaborative browser based IDE - Replit

ECAC

Portal do Contribuinte

Pericia

Tabela de honorários de perito - cálculo instantâneo - Manual de Perícias - Seja Perito Judicial | Maior sistema de conteúdo sobre perícia judicial
Perícia em Licenciamento de Software: Boas Práticas e fabricantes que mais demandam » Lopes Perícias
A profissão do futuro: Como ser um perito forense digital ou perito em informática e faturar na carreira? | Jusbrasil
AR Soluti

OiPlay
Oi

PD

Empregos

MeiraFernandes

Outros favoritos

Adiados - ecacarva@gmail.com - Gmail
Inte_r net___B_ankiNg____CAiXA
Hello, gostariamos de te entender melhor - ecacarva@gmail.com - Gmail
RPubs - Exercício - R Markdown - Regressão

INSS

New Folder

Install MySQL 8 on CentOS 8 - kifarunix.com
Firewall iptables no Linux - Parte 03: Exemplos de regras - Bóson Treinamentos em Tecnologia
Avaliações de Geografia sobre Paisagens - Atividades Pedagógicas
ATIVIDADES DE ALFABETIZAÇÃO: AUTODITADOS PARA IMPRIMIR - Mistura de Alegria
https://calendar.google.com/calendar/r/week?tab=mc&pli=1
Error ‹ WordPress.com — WordPress
Home - Journal Article Publishing Support Center

Exame

The fastest way to compile your Table 1?
CURSO MOODLE #4 -INSTALAÇÃO DE TEMAS EXTERNOS NO MOODLE - YouTube
Customização Moodle - Meu Dashboard como deve ser! - YouTube
Como configurar plugin WP Mail SMTP no WordPress – Suporte HostGator
Como usar o Moodle - YouTube

00.Comida

30 das mais fáceis receitas para fazer no Dia dos Pais • Cozinha Legal

My No-IP :: Dashboard
Richard Paap: Quantitative Models in Marketing Research
Sistema de Ensino à Distância: Administração: Plugins: Atividades: Chat
Os Móveis mais desejados do País - Viva Bem - Viva Parma
SciELO - Saúde Pública - Structural equation modeling in epidemiology Structural equation modeling in epidemiology
Um tutorial completo sobre modelagem baseada em árvores de decisão (códigos R e Python) – Vooo – Insights
Counterfactual Reasoning for Process Optimization Using Structural Causal Models | SpringerLink

SQL

(1) SQL Joins Tutorial for Beginners - Inner Join, Left Join, Right Join, Full Outer Join - YouTube

Um estudo sobre modelos de aprendizagem baseados em árvores com desafio do Kaggle
01.Understanding random forests with randomForestExplainer

CARET

A basic tutorial of caret: the machine learning package in R
RPubs - Random Forest Tunning in Caret
5 Model Training and Tuning | The caret Package
Caret Package - A Complete Guide to Build Machine Learning in R
A Brief Introduction to caret — Kimberly Coffey, Ph.D.

TidyModels

Tidymodels: tidy machine learning in R
Fulljoin: Prevendo a qualidade de vinhos com o tidymodels

Tidyverse

Books - Statistical Inference via Data Science
Tidyverse

machine learning - Binary classification with strongly unbalanced classes - Cross Validated

New Folder

Nova pasta

New Folder

Nova pasta

Biblio

ABCDWIKI Español
FOLIO – O Futuro das Bibliotecas é aberto – Software livre para bibliotecas – Bibliotecários Sem Fronteiras
ABCD Downloads Page
Software livre para biblioteca: quem usa o ABCD no Brasil?
ABCD - Software para gestão de biblioteca - L2O Consultoria
Home
Biblivre 5 - Baixar
Caixa de entrada
ABCDLibrary » Versões ABCD

IJCAI2020

Submissão
Conference Management Toolkit - Submission details
IJCAI-PRICAI 2020

Viagem

Dall'Onder Grande Hotel
Trem Maria Fumaça + Epopeia + Degustações | Início às 17:30h - TudoPass
Rotas Brasil - Cálculo de pedágios, tempo, distância e custos
Hoteis 10
GPS IGO Nova Central Nissan March e Versa 16/17 18 como instalar atualização 2020 - YouTube
Google Drive começam em R$ 7 por mês - Pesquisa Google

VBox

Como solucionar o erro de dependência Libpng12-0 | SempreUPdate

WordPress

5 Melhores Plugins de Tradução WordPress
Tudo para Wordpress e Mautic - Hospedagem, Cursos e Ferramentas - Hostnet – Hospedagem de Sites com Domínio Grátis
Livre Sat

Livros

L

Introdução à análise Bayesiana (com R) - 9788561797102 - Livros na Amazon Brasil
Introdução à Análise de Dados Categóricos com Aplicações: Suely Ruiz Giolo: Amazon.com.br: Livros

Uningá - Site

Uningá — WordPress
https://uninga.br/cursos/mestrado/
Linux Mint - Community
WordPress › Setup Configuration File
A Carregar: 14% — Convertio
Bem-vindo ao Programa de Pós-Graduação em Ciências da Saúde - Mestrado e Doutorado — Programa de Pós-Graduação em Ciências da Saúde
I:ntern.et.... ____----___BAnki____ng____.......CAIXa
Site multilíngue com WordPress - YouTube
"florestas aleatórias" Randomforest R - Google Acadêmico
cPanel File Manager v3

Cantos Gregorianos Áudio - Arautos do Evangelho

CENTOS

Comandos CentOS 7 » SECNET
Squid no Cent Os 7 bloqueia todos os sites https [Iniciantes no Linux]

Aulas Claretiano

Atividades de Artes 1 ano – Educação e Transformação
ATIVIDADES DE CIÊNCIAS PARA O 4º ANO - ENSINO FUNDAMENTAL
50 Atividades de Ciências 5º Ano Ensino Fundamental para Imprimir - Online Cursos Gratuitos
EDUCAÇÃO FÍSICA - 45 ATIVIDADES EXERCÍCIOS DESENHOS COLORIR IMPRIMIR - PORTAL ESCOLA
(357) Educação Física: Atividade Diferenciada - Ensino Fundamental: Aula de 07/06/2017 - Colégio Messina - YouTube
ATIVIDADES DE GEOGRAFIA PARA O 5º ANO — SÓ ESCOLA
aula de história 5o ano fundamental - Pesquisa Google
Arquivos atividade de interpretação de texto 5º ano - Acessaber
(357) Jogo oito cones - Educação Física - YouTube
Atividades de Geografia 4º Ano
Atividades de Geografia para 5° ano - ESPAÇO EDUCAR
Atividades de História 4º Ano
Atividades de Língua Portuguesa: Poema e interpretação (para 3º ou 4º Ano)
Atividades 5º Ano de Língua Portuguesa: Entendendo as ideias principais de um texto
Atividades de Língua Portuguesa: Correspondências irregulares e regulares
Atividades de matemática do 4º ano [21 para imprimir]
Atividades de Matemática 4º ano - Elemento neutro - Escola Educação
Atividades de Matemática para 5º Ano - Reforço — SÓ ESCOLA
aula de natureza e sociedade 5o ano fundamental - Pesquisa Google
Atividades de Produção de Texto - 2º ao 5º ano — SÓ ESCOLA
26 excelentes Atividades de Artes para Imprimir: Ensino fundamental | Max Dicas
Atividades de Matemática - 5º ano - Para Imprimir
Blog Professor Zezinho : Atividades sobre sistema respiratório para 5° ano ou 4ª série - Atividades sobre respiração, partes do aparelho respiratório, pulmões etc
Atividades de Ciências para Educação Infantil - 1º ano
matematica ensino infantil 2 - Pesquisa Google
Atividades de matemática para Educação Infantil - Atividades Pedagógicas
ALFABETIZAÇÃO DIVERTIDA: FÁBULA: O MACACO E O COELHO
Atividades identidade e autonomia para Educação Infantil
Atividades para o maternal - natureza e sociedade (29).jpg (JPEG Image, 638 × 850 pixels)
(360) D-13 - Identidade e Autonomia - YouTube

Hardware Test

16 commands to check hardware information on Linux – BinaryTides
10 Useful Commands to Collect System and Hardware Information in Linux
Check Hardware Information On Linux via Command Line
How to see CPU temperature on CentOS 7 and RedHat Enterprise Linux 7 - nixCraft
command to check hardware
Unix Toolbox
Unix Toolbox
How to obtain information about network devices and their configuration on Linux - LinuxConfig.org
Testes simples de desempenho de hardware em sistemas Linux

Exames

Emissão Laudo
EAD
Resultados Online

no-ip

Open Port Check Tool - Test Port Forwarding on Your Router
MeuIP - Qual é o meu ip?

Oi

Oi | Combo, TV, Celular, Internet, Fixo, Recarga
EBPP | Login

Correlação

e-Aulas da USP :: Correlação de Spearman (SigmaPlot 12.0)

PCA

PCA Analysis in R - DataCamp
RPubs - Análise de Componentes Principais
Capítulo 6 Redução de Dimensionalidade | Estatística Avançada
Tutorial R | Analise de correspondência múltipla ACM ou MCA no R - YouTube
A análise de componentes principais pode ser aplicada a conjuntos de dados que contêm uma mistura de variáveis ​​contínuas e categóricas?
FactoMineR: Exploratory Multivariate Data Analysis with R

Regressão Logística

RPubs - Regressão Logística em R
Regressão Logística no R - YouTube
(2) Regressão Logística Binária no SPSS (Aula 13 - Parte 2) - YouTube
(2) Binary logistic regression using R: Analysis and interpretation - YouTube
Modelo de Regressão Logística Binária com R prof. José Rodrigo de Moraes/UFF | Estatística é com R!
Capítulo 7 Regressão Logística | Software R: curso avançado
Interpreting the concordance statistic of a logistic regression model: relation to the variance and odds ratio of a continuous explanatory variable | BMC Medical Research Methodology | Full Text
(99) StatQuest: Odds and Log(Odds), Clearly Explained!!! - YouTube
(99) Interpretação de OR, RR, ... Parte III: exemplos de interpretação de artigos científicos - YouTube

Parma

Escritório Linha Clássica - Parma Móveis

ML

Explaining Black-Box Machine Learning Models - Code Part 1: tabular data + caret + iml
Hands-On Machine Learning with R
STHDA - Home
Machine Learning Mastery With R

R Packages

Reading and Writing Excel Files With R Using readxl and writexl

MCA

(1) Análise de Correspondência Múltipla (MCA) - YouTube
MCA - Multiple Correspondence Analysis in R: Essentials - Articles - STHDA
MCA in R Using FactoMineR: Quick Scripts and Videos - Articles - STHDA
análise de correspondência múltipla no R - Pesquisa Google

Consagração

Curso de Consagração a Nossa Senhora pelo método de São Luís Grignion de Montfort – Formação Católica Arautos do Evangelho

Zero inflado

Zero Inflado | Recologia
GLM com distribuição de Poisson e zeros inflados | Recologia

X DRONE HD | high-end drone at an affordable price
Campaigns | Mailchimp

COVID19

COVID-19 no Brasil
Visualize the Pandemic with R #COVID-19 - Towards Data Science
nCov2019 for studying COVID-19 coronavirus outbreak
Download de dados de COVID-19 no Brasil • coronabr
Testando o pacote CoronaBr [R] :: Ícaro Agostino — Forecasting, Simulção e Otimização
Portal da Transparência - Registro Civil
Covid-19: Brasil registra 1.185 mortes em 24h; total ultrapassa 52 mil

phpMyAdmin
saocamilo
RIS - Página principal
atos dos apóstolos 2 - Pesquisa Google
SOGo
google - Pesquisa Google
Liturgia Diária
Google
Entrada (925) - msstabor@gmail.com - Gmail
Congregatio - Identificação
Identificação | Fotos Chez Nous
Fazer login nas Contas do Google
Google Tradutor
Gmail
https://maringa.arautos.org/
Arquivocheznous :: Login
arquidiocese de maringa - Pesquisa Google
Arquidiocese de Maringá - Paraná
about:blank
Ex-Arautos – Minas Gerais – Adeptos da verdade
Email – Mário Sérgio Sperche – Outlook
MEGA ↓ 75%
Arautos: Ver, Julgar e Agir! -
Crase antes de pronomes - Português
Banco Bradesco | Pessoa Física, Exclusive, Prime e Private
Exorcismo Arautos -
Grupo São Camilo - Medicina Diagnóstica
Colégio Arautos do Evangelho | Site oficial do Colégio Arautos do Evangelho
RIS - Downloads
Arautos Véritas - Toda a verdade sobre os Arautos
Entrar na conta da Microsoft
Congregatio - Identificação
Acessar ‹ — WordPress
Imagem Escultura São José Rei 22cm - Nova Lumen | Presentes e Artigos
Nossa Senhora do Rosário
Arautos em Missão - Página inicial | Facebook
Portal Nacional de Saúde :: Hotsite Laboratório Unimed Campos :: Exames e Procedimentos
TVArautos - YouTube
WhatsApp
EAD - Colégio Arautos do Evangelho de Maringá
Painel
Internet: o Reino do Fakenews.... - pe.mariosergioep@gmail.com - Gmail
2srt2ass
Meet

Projeto Dupla

Data Scientist JR
ENV-Desenvolvedor Qlikview ou Qlik Sense; - Vagas Floripa - Florianópolis
Instruções por voz gratuitas, informações de trânsito e App de Navegação por GPS feito pelo Waze.

Multicollinearity

The GoodmanKruskal package: Measuring association between categorical variables

Oi Mais Empresas
Google Tradutor
suportecheznous.org: Acesso ao site
lattent

Nova pasta

WhatsApp
SumUp | A máquina de cartão número 1
Carta Grupo VIP – VCMD – Por Professor Ezequiel Mendes

Paula Abrante - Referencias

SciELO - Brasil - Projeção populacional, mudanças climáticas e efeitos econômicos: uma avaliação a partir de blocos econômicos agrícolas Projeção populacional, mudanças climáticas e efeitos econômicos: uma avaliação a partir de blocos econômicos agrícolas
World Population Prospects 2019: Methodology of the United Nations population estimates and projections - WPP2019_Methodology.pdf
mysql access denied for user 'root' - Pesquisa Google